Refactored patterns into separate modules
Separated the patterns for each of the currently implemented join
patterns into separate files.
This should make future development considerably easier.
